Create utils file for creating sql tables
