+++ components/clp-py-utils/clp_py_utils/initialize-clp-metadata-db.py
@@ -13,6 +13,7 @@ from clp_py_utils.clp_config import (
     DATASETS_TABLE_SUFFIX,
 )
 from clp_py_utils.core import read_yaml_config_file
+from clp_py_utils.sql_table_schema_utils import create_archives_table, create_column_metadata_table
 
 # Setup logging
 # Create logger
@@ -40,23 +41,7 @@ def main(argv):
         with closing(sql_adapter.create_connection(True)) as metadata_db, closing(
             metadata_db.cursor(dictionary=True)
         ) as metadata_db_cursor:
-            metadata_db_cursor.execute(
-                f"""
-                CREATE TABLE IF NOT EXISTS `{table_prefix}{ARCHIVES_TABLE_SUFFIX}` (
-                    `pagination_id` BIGINT unsigned NOT NULL AUTO_INCREMENT,
-                    `id` VARCHAR(64) NOT NULL,
-                    `begin_timestamp` BIGINT NOT NULL,
-                    `end_timestamp` BIGINT NOT NULL,
-                    `uncompressed_size` BIGINT NOT NULL,
-                    `size` BIGINT NOT NULL,
-                    `creator_id` VARCHAR(64) NOT NULL,
-                    `creation_ix` INT NOT NULL,
-                    KEY `archives_creation_order` (`creator_id`,`creation_ix`) USING BTREE,
-                    UNIQUE KEY `archive_id` (`id`) USING BTREE,
-                    PRIMARY KEY (`pagination_id`)
-                )
-                """
-            )
+            create_archives_table(metadata_db_cursor, f"{table_prefix}{ARCHIVES_TABLE_SUFFIX}")
 
             metadata_db_cursor.execute(
                 f"""
@@ -109,14 +94,8 @@ def main(argv):
                 """
             )
 
-            metadata_db_cursor.execute(
-                f"""
-                CREATE TABLE IF NOT EXISTS `{table_prefix}default_{COLUMN_METADATA_TABLE_SUFFIX}` (
-                    `name` VARCHAR(512) NOT NULL,
-                    `type` TINYINT NOT NULL,
-                    PRIMARY KEY (`name`, `type`)
-                )
-                """
+            create_column_metadata_table(
+                metadata_db_cursor, f"{table_prefix}default_{COLUMN_METADATA_TABLE_SUFFIX}"
             )
 
             metadata_db.commit()
