+++ components/clp-py-utils/clp_py_utils/sql_table_schema_utils.py
@@ -1,0 +1,30 @@
+def create_archives_table(db_cursor, table_name: str) -> None:
+    db_cursor.execute(
+        f"""
+        CREATE TABLE IF NOT EXISTS `{table_name}` (
+            `pagination_id` BIGINT unsigned NOT NULL AUTO_INCREMENT,
+            `id` VARCHAR(64) NOT NULL,
+            `begin_timestamp` BIGINT NOT NULL,
+            `end_timestamp` BIGINT NOT NULL,
+            `uncompressed_size` BIGINT NOT NULL,
+            `size` BIGINT NOT NULL,
+            `creator_id` VARCHAR(64) NOT NULL,
+            `creation_ix` INT NOT NULL,
+            KEY `archives_creation_order` (`creator_id`,`creation_ix`) USING BTREE,
+            UNIQUE KEY `archive_id` (`id`) USING BTREE,
+            PRIMARY KEY (`pagination_id`)
+        )
+        """
+    )
+
+
+def create_column_metadata_table(db_cursor, table_name: str) -> None:
+    db_cursor.execute(
+        f"""
+        CREATE TABLE IF NOT EXISTS `{table_name}` (
+            `name` VARCHAR(512) NOT NULL,
+            `type` TINYINT NOT NULL,
+            PRIMARY KEY (`name`, `type`)
+        )
+        """
+    )
